# [Go] Reorganize code and write tests for profiling