--- a/go/go.mod
+++ b/go/go.mod
@@ -1,3 +1,3 @@
 module g.io/related
 
-go 1.21.3
+go 1.21.4