Noticed a decent perf improvement even though the algo wasn't changed. `20ms -> 16ms`

Main: `19.11 ms | 281.24 ms | 2.48 s`

This PR:
```
Go:

        Processing time (w/o IO): 17.359155ms
        total: 0.04s memory: 13056k
        Processing time (w/o IO): 16.517505ms
        total: 0.04s memory: 13180k
        Processing time (w/o IO): 16.412455ms
        total: 0.04s memory: 12928k
        Processing time (w/o IO): 16.441164ms
        total: 0.04s memory: 13184k
        Processing time (w/o IO): 17.691736ms
        total: 0.04s memory: 13440k
        Processing time (w/o IO): 16.365855ms
        total: 0.04s memory: 13020k
        Processing time (w/o IO): 16.935075ms
        total: 0.04s memory: 13044k
        Processing time (w/o IO): 17.970876ms
        total: 0.04s memory: 13044k
        Processing time (w/o IO): 16.967825ms
        total: 0.04s memory: 12928k
        Processing time (w/o IO): 16.291865ms
        total: 0.04s memory: 13440k

Go:

        Processing time (w/o IO): 245.080777ms
        total: 0.35s memory: 43960k
        Processing time (w/o IO): 238.42926ms
        total: 0.35s memory: 44608k
        Processing time (w/o IO): 240.839094ms
        total: 0.35s memory: 44060k

Go:

        Processing time (w/o IO): 2.110623954s
        total: 2.49s memory: 161400k
        Processing time (w/o IO): 2.104082079s
        total: 2.48s memory: 161628k
        Processing time (w/o IO): 2.131778195s
        total: 2.50s memory: 161008k
```